postgres_to_es/requirements.txt
backoff==1.10.0
elasticsearch>=7.0.0,<8.0.0
pydantic==1.6.1
psycopg2-binary==2.8.6
redis==3.5.3
python-dotenv==0.14.0
requests